refactor `getCursorFromHashParams`, `updateViewHashParams`, and `updateQueryHashParams` to urlHash
