+++ src/stores/viewStore/createViewPageSlice.ts
@@ -13,7 +13,7 @@ import {
     NavigationAction,
 } from "../../utils/actions";
 import {clamp} from "../../utils/math";
-import {updateWindowUrlHashParams} from "../../utils/url.ts";
+import {updateWindowUrlHashParams} from "../../utils/url";
 import useLogFileManagerStore from "../logFileManagerProxyStore";
 import useLogFileStore from "../logFileStore";
 import {handleErrorWithNotification} from "../notificationStore";
